Directory shuffle.
A standalone repository for Cycles has been created at
git.blender.org ( developer.blender.org ). These changes are in
preparation of using that standalone repository as submodule in
one way or another. I'll also be making some changes to how the
work-flow in this repository is going to be handled. Most notably
that master will be more like what bindep is now, and a separate
branch for project file customisations as necessary for RhinoCycles
development.